ttrader/tests/test_account.py
import sqlite3
import os
import unittest
from app.orm import ORM
from app import Account, Position, Trade
from data.seed import seed
from data.schema import schema

DIR = os.path.dirname(__file__)
DBFILENAME = "_test.db"
DBPATH = os.path.join(DIR, DBFILENAME)

""" setting ORM.dbpath changes the db for all classes inheriting from it """
ORM.dbpath = DBPATH


class TestAccount(unittest.TestCase):
    def setUp(self):
        """ the setup method must be called setup """
        schema(DBPATH)
        seed(DBPATH)

    def tearDown(self):
        """ the tear down method must be called tearDown """
        os.remove(DBPATH)

    def test_save_and_pk_load(self):
        user = Account(username="Samuel")
        user.save()
        self.assertIsInstance(user.pk, int, 'save sets pk')
        
        pk = user.pk
        same_user = Account.one_from_pk(pk)

        self.assertIsInstance(same_user, Account, "one_from_pk loads an Account object")

        self.assertEqual(same_user.username, "Samuel", "save creates database row")
        same_user.username = "Sam"
        same_user.save()

        same_again = Account.one_from_pk(pk)

        self.assertEqual(same_again.username, "Sam", "save updates an existing row")

    def test_get_positions(self):
        user = Account.one_from_pk(1)
        positions = user.get_positions()
        self.assertIsInstance(positions, list, "get positions returns a list")
        self.assertIsInstance(positions[0], Position, "list contains Position objects")

    def test_get_position_for(self):
        pass

    def test_get_trades(self):
        pass

    def test_get_trades_for(self):
        pass

    def test_login(self):
        mike_bloom = Account.login('mike_bloom', 'password')
       
        self.assertIsNotNone(mike_bloom, "account & password find row")

        self.assertIsInstance(mike_bloom, Account, "login returns an account")

    def test_buy(self):
        pass

    def test_sell(self):
        pass
